requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
altgraph==0.17.4
amqp==5.2.0
anyio==4.1.0
async-timeout==4.0.3
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.95.1
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-resources==6.1.1
kombu==5.3.4
limits==3.7.0
motor==3.1.2
multidict==6.0.4
orjson==3.8.12
packaging==23.2
passlib==1.7.4
pefile==2023.2.7
persiantools==3.0.1
pillow==10.4.0
prompt-toolkit==3.0.41
pyasn1==0.5.1
pydantic==1.10.13
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pymongo==4.6.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
redis==4.5.4
rsa==4.9
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.8.0
uvicorn==0.22.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
wrapt==1.16.0
yarl==1.9.3
zipp==3.16.2